[FD] Move MobileHelpToSaveConfig and related inside the mobilehelptosave package
Looks like it was initially created in the wrong package by mistake.
