+++ app/uk/gov/hmrc/mobilehelptosave/connectors/HelpToSaveConnector.scala
@@ -25,9 +25,9 @@ import javax.inject.{Inject, Singleton}
 import org.joda.time.{LocalDate, YearMonth}
 import play.api.LoggerLike
 import play.api.libs.json.{JsValue, Json, OFormat}
-import uk.gov.hmrc.config.HelpToSaveConnectorConfig
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http._
+import uk.gov.hmrc.mobilehelptosave.config.HelpToSaveConnectorConfig
 import uk.gov.hmrc.mobilehelptosave.config.ScalaUriConfig.config
 import uk.gov.hmrc.mobilehelptosave.config.SystemId.SystemId
 import uk.gov.hmrc.mobilehelptosave.domain._
